Update dependency usage in Authentication Service
This modification enhances the IAuthenticationService and TestUserAuthenticationService by incorporating the System namespace. The change involves removing the dependency on Unity.Plastic.Newtonsoft.Json.Serialization and replacing it with the System namespace. This adjustment streamlines the code and improves the maintainability of the Authentication Service.

In the IAuthenticationService interface, the System namespace is introduced to enhance its functionality. This update allows the interface to leverage the System namespace's features while eliminating the need for the previously used dependency.

Similarly, the TestUserAuthenticationService class also benefits from the inclusion of the System namespace. By removing the dependency on Unity.Plastic.Newtonsoft.Json.Serialization and incorporating the System namespace, the class becomes more efficient and easier to maintain. This change positively impacts the overall performance of the Authentication Service.